+++ src/app/product/product.controller.js
@@ -57,7 +57,7 @@
 
         function viewProductHistory () {
             vm.viewProductHistoryInstance = $uibModal.open({
-                templateUrl: 'app/product/product_history.html',
+                templateUrl: 'app/product/history/history.html',
                 controller: 'ProductHistoryController',
                 controllerAs: 'vm',
                 animation: false,