+++ src/app/admin/components/surveillance/inspect.controller.spec.js
@@ -1,0 +1,56 @@
+(function () {
+    'use strict';
+
+    describe('admin.InspectSurveillanceController.controller', function () {
+        var vm, scope, $log, $q, commonService, utilService, Mock;
+
+        beforeEach(function () {
+            module('chpl.mock', 'chpl.admin', function ($provide) {
+                $provide.decorator('commonService', function ($delegate) {
+                    $delegate.getSurveillanceLookups = jasmine.createSpy('getSurveillanceLookups');
+                    return $delegate;
+                });
+                $provide.decorator('utilService', function ($delegate) {
+                    $delegate.sortRequirements = jasmine.createSpy('sortRequirements');
+                    return $delegate;
+                });
+            });
+
+            inject(function ($controller, _$log_, _$q_, $rootScope, _Mock_, _commonService_, _utilService_) {
+                $log = _$log_;
+                $q = _$q_;
+                commonService = _commonService_;
+                commonService.getSurveillanceLookups.and.returnValue($q.when([]));
+                utilService = _utilService_;
+                utilService.sortRequirements.and.returnValue(1);
+                Mock = _Mock_;
+
+                scope = $rootScope.$new();
+                vm = $controller('SurveillanceInspectController', {
+                    surveillance: Mock.surveillances[0],
+                    $uibModalInstance: Mock.modalInstance,
+                    $scope: scope,
+                });
+                scope.$digest();
+            });
+        });
+
+        afterEach(function () {
+            if ($log.debug.logs.length > 0) {
+                //console.debug('\n Debug: ' + $log.debug.logs.join('\n Debug: '));
+            }
+        });
+
+        describe('housekeeping', function () {
+            it('should exist', function () {
+                expect(vm).toBeDefined();
+            });
+
+            it('should have a way to close the modal', function () {
+                expect(vm.cancel).toBeDefined();
+                vm.cancel();
+                expect(Mock.modalInstance.dismiss).toHaveBeenCalled();
+            });
+        });
+    });
+})();