+++ RELEASE_NOTES.md
@@ -23,6 +23,7 @@ _Date TBD_
     * All retired ACBs
 * Changed admin Notifications section to just "Subscription Management"
 * Added "API Documentation" collection
+* Added Listing History explanatory text
 
 ---