+++ RELEASE_NOTES.md
@@ -24,6 +24,7 @@ _Date TBD_
 * Changed admin Notifications section to just "Subscription Management"
 * Added "API Documentation" collection
 * Added Listing History explanatory text
+* Updated summary description of surveillance result
 
 ---