+++ RELEASE_NOTES.md
@@ -23,6 +23,8 @@ _Date TBD_
     * All retired ACBs
 * Changed admin Notifications section to just "Subscription Management"
 * Added "API Documentation" collection
+* Added Listing History explanatory text
+* Updated summary description of surveillance result
 
 ---