+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+OCD-1552: Allow scrolling on CMS ID & Compare widgets
+OCD-1463: Ellipsis long names on CMS ID & Compare widgets
+
+---
+
 ## Version TBD
 _Date TBD_